Create decks/remove.rs
